% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_inputs.R
\name{add_days}
\alias{add_days}
\title{Add a specified number of days to a date}
\usage{
add_days(yyyymmdd, n)
}
\arguments{
\item{yyyymmdd}{date in YYYYMMDD format}

\item{n}{number of days to add to date}
}
\value{
yyyymmdd + n
}
\description{
Add a specified number of days to a date
}
