% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_inputs.R
\name{add_months}
\alias{add_months}
\title{Add a specified number of months to a date}
\usage{
add_months(yyyymm, n)
}
\arguments{
\item{yyyymm}{date in YYYYMM format}

\item{n}{number of months to add to date}
}
\value{
yyyymm + n
}
\description{
Add a specified number of months to a date
}
