% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_inputs.R
\name{add_years}
\alias{add_years}
\title{Add a specified number of years to a year}
\usage{
add_years(yyyy, n)
}
\arguments{
\item{yyyy}{year in YYYY format}

\item{n}{number of years to add}
}
\value{
yyyy + n
}
\description{
Add a specified number of years to a year
}
