% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_to_bin.R
\name{assign_to_bin}
\alias{assign_to_bin}
\title{Assign a value to one of several bins.}
\usage{
assign_to_bin(vals, bins)
}
\arguments{
\item{vals}{Values to be assigned to a bin}

\item{bins}{Bin values}
}
\value{
a bin corresponding to each of \code{vals}
}
\description{
Values less than the smallest bin value are assigned to that
value. Values greater than the largest bin value are assigned
to that value.
}
