% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_blockwise_apply.R
\name{block_apply}
\alias{block_apply}
\title{Apply a function to each cell in a GDALRasterDataset, using blockwise I/O}
\usage{
block_apply(rast, rast_out, fun, band = 1)
}
\arguments{
\item{rast}{A GDALRasterDataset from which values should be read}

\item{rast_out}{A GDALRasterDataset to which values should be written}

\item{fun}{A function to apply to each cell in \code{rast}}

\item{band}{The band of \code{rast} to read}
}
\description{
Apply a function to each cell in a GDALRasterDataset, using blockwise I/O
}
