% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_to_integer.R
\name{can_coerce_to_integer}
\alias{can_coerce_to_integer}
\title{Test if values can be stored as integers without loss}
\usage{
can_coerce_to_integer(vals)
}
\arguments{
\item{vals}{a vector of values to test}
}
\value{
\code{TRUE} if \code{vals} can be cast to integers without loss
}
\description{
Test if values can be stored as integers without loss
}
