% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_write.R
\name{can_write}
\alias{can_write}
\title{Determine if a file can be opened for writing}
\usage{
can_write(filename)
}
\arguments{
\item{filename}{filename to test for write access}
}
\value{
TRUE if the file can be opened for writing,
        FALSE otherwise
}
\description{
If the file exists, check if we have write
permissions on it.
}
\details{
If the file does not exist, try creating a file
of that name, and then remove it.
}
