% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{check_coordinate_variable}
\alias{check_coordinate_variable}
\title{Validate coordinate variables}
\usage{
check_coordinate_variable(ncout, varname, values)
}
\arguments{
\item{ncout}{a netCDF file opened for writing}

\item{varname}{the name of a coordinate variable}

\item{values}{expected values of the coodinate variable}
}
\description{
Validate coordinate variables
}
