% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_units.R
\name{check_units}
\alias{check_units}
\title{Check the units of data read by \code{read_vars_from_cdf}}
\usage{
check_units(data, var, expected_units, fname)
}
\arguments{
\item{data}{object returned by \code{read_vars_from_cdf}}

\item{var}{name of variable to check}

\item{expected_units}{expected units of variable (e.g., 'mm')}

\item{fname}{filename from which variable was read (used for
creating an error message only)}
}
\value{
nothing
}
\description{
If the units are undefined, a warning will be issued.
If the units are incorrect, a the program will exit.
}
