% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{check_values_exist_in_dimension}
\alias{check_values_exist_in_dimension}
\title{Validate extra dimensions}
\usage{
check_values_exist_in_dimension(ncout, vars)
}
\arguments{
\item{ncout}{a netCDF file opened for writing}

\item{vars}{list containing values of any named dimension variables}
}
\description{
Validate extra dimensions
}
