% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{check_var_list}
\alias{check_var_list}
\title{Get a list of variables that can be read from a netCDF file}
\usage{
check_var_list(cdf, vars)
}
\arguments{
\item{cdf}{ncdf4 file handle}

\item{vars}{a (possibly-empty) list of var objects to read from \code{cdf}}
}
\value{
a list of \code{var} objects.
}
\description{
A list of requested variable names can be provided.
}
