% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combos}
\alias{combos}
\title{Return a data frame with each combination of the supplied inputs}
\usage{
combos(...)
}
\arguments{
\item{...}{one or more named arguments}
}
\value{
a data frame
}
\description{
Return a data frame with each combination of the supplied inputs
}
