% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{create_char_dimension_variable}
\alias{create_char_dimension_variable}
\title{Create a character-type dimension variable}
\usage{
create_char_dimension_variable(dim, varname, vals)
}
\arguments{
\item{dim}{\code{ncdim} handle}

\item{varname}{name of dimension/variable}

\item{vals}{values of dimension}
}
\description{
Create a character-type dimension variable
}
