% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_blockwise_apply.R
\name{create_tiff_like}
\alias{create_tiff_like}
\title{Create a new GeoTIFF with the same dimensions/projection the input}
\usage{
create_tiff_like(rast, filename_out, datatype)
}
\arguments{
\item{rast}{a GDALRasterDataset}

\item{filename_out}{filename of output GeoTIFF}

\item{datatype}{datatype of values in output GeoTIFF}
}
\value{
a GDALRasterDataset pointing to the output GeoTIFF
}
\description{
Create a new GeoTIFF with the same dimensions/projection the input
}
