% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{create_var}
\alias{create_var}
\title{Create a ncvar4 object}
\usage{
create_var(dims, varname, vals, prec, compress)
}
\arguments{
\item{dims}{list of \code{ncdim4} objects}

\item{varname}{name of the variable}

\item{vals}{values for the variable}

\item{prec}{argument passed to write_vars_to_cdf}

\item{compress}{should compression be enabled for the new variable?}
}
\value{
a \code{ncvar4} object for the variable
}
\description{
Create a ncvar4 object
}
