% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_to_matrices.R
\name{cube_to_matrices}
\alias{cube_to_matrices}
\title{Convert a 3D array to a list of matrices}
\usage{
cube_to_matrices(cube)
}
\arguments{
\item{cube}{a 3-dimensional array}
}
\value{
a list of matrices
}
\description{
The dimnames of the 3rd dimension will be
used to name the elements of the returned
list.
}
