% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_precl.R
\name{curl_range}
\alias{curl_range}
\title{Download a range of bytes from a URL using curl}
\usage{
curl_range(url, start, stop, fname, timeout = NULL)
}
\arguments{
\item{url}{url to access}

\item{start}{first byte to read (0-indexed, inclusive)}

\item{stop}{last byte to read (0-indexed, inclusive)}

\item{fname}{output filename}

\item{timeout}{timeout in seconds}
}
\description{
Download a range of bytes from a URL using curl
}
