% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_inputs.R
\name{date_range}
\alias{date_range}
\title{Expand a range of dates}
\usage{
date_range(start, stop, step)
}
\arguments{
\item{start}{start of range in YYYYMM format}

\item{stop}{end of range in YYYYMM format}

\item{step}{number of months between range
elements.}
}
\value{
a vector of dates in YYYYMM format
}
\description{
Expand a range of dates
}
