% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\docType{data}
\name{default_netcdf_nodata}
\alias{default_netcdf_nodata}
\title{Default NODATA values for various data types}
\format{
An object of class \code{list} of length 7.
}
\usage{
default_netcdf_nodata
}
\description{
Default NODATA values for various data types
}
\keyword{datasets}
