% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_with_message.R
\name{die_with_message}
\alias{die_with_message}
\title{Display a message and exit the program}
\usage{
die_with_message(...)
}
\arguments{
\item{...}{A list that will be concatenated into an
error message}
}
\description{
All arguments will be concatenated together.
Program will exit with status=1
}
