% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{error}
\alias{error}
\title{Generate a log message at the \code{ERROR} level}
\usage{
error(...)
}
\arguments{
\item{...}{Any number of string-convertible objects}
}
\description{
Generate a log message at the \code{ERROR} level
}
