% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_inputs.R
\name{expand_dates}
\alias{expand_dates}
\title{Expand any date ranges encoded in a filename}
\usage{
expand_dates(fname)
}
\arguments{
\item{fname}{file name or glob, possibly containing
one or more date ranges between square
brackets}
}
\value{
a vector of filenames or globs
}
\description{
Expand any date ranges encoded in a filename
}
