% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_inputs.R
\name{expand_inputs}
\alias{expand_inputs}
\title{Expand a list of inputs using globbing and date range
expansion, returning a single list of filenames.}
\usage{
expand_inputs(raw_inputs, check_exists = TRUE)
}
\arguments{
\item{raw_inputs}{a list of file path globs}

\item{check_exists}{should we throw and error if
a specified file does not exist,
or if a glob returns no matches?}
}
\value{
a list of filenames
}
\description{
Glob expansion is performed using a system call.
}
\details{
Date range expansion is performed by including
a date range in square brackets, e.g. \code{results_[201201:201203].nc}, or
\code{results[195004:200904:12]}, where the third value
indicates the number of months between successive
dates in the range.

The list of filenames may contain duplicates.

If any supplied filename, date range, or glob does
not match an existing file, this function will
throw an error, unless \code{check_exists} is set to
\code{FALSE}.
}
