% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{fatal}
\alias{fatal}
\title{Generate a log message at the \code{FATAL} level}
\usage{
fatal(...)
}
\arguments{
\item{...}{Any number of string-convertible objects}
}
\description{
Generate a log message at the \code{FATAL} level
}
