% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{find_count}
\alias{find_count}
\title{Compute the count vector for specified valuws of dimensions}
\usage{
find_count(real_dims, dim_values)
}
\arguments{
\item{real_dims}{a list of dimensions}

\item{dim_values}{a list whose keys represent dimension names and values
represent values along those dimensions}
}
\description{
Compute the count vector for specified valuws of dimensions
}
