% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{find_offset}
\alias{find_offset}
\title{Compute the offset vector for specified values of dimensions}
\usage{
find_offset(cdf, real_dims, dim_values)
}
\arguments{
\item{cdf}{a ncdf4 file handle}

\item{real_dims}{a list of dimensions}

\item{dim_values}{a list whose keys represent dimension names and values
represent values along those dimensions}
}
\description{
Compute the offset vector for specified values of dimensions
}
