% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_arr.R
\name{flatten_arr}
\alias{flatten_arr}
\title{Reduce the dimensions of a 3-dimensional array}
\usage{
flatten_arr(arr)
}
\arguments{
\item{arr}{array to flatten}
}
\description{
The first two dimensions will be combined, while the third dimension will be preserved.
}
