% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{flip_first_two_dims}
\alias{flip_first_two_dims}
\title{Flip the first two dimensions of a multidimensional array}
\usage{
flip_first_two_dims(arr)
}
\arguments{
\item{arr}{an array whose dimensions should be flipped}
}
\value{
the transformed array
}
\description{
Flip the first two dimensions of a multidimensional array
}
