% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmme.R
\name{forecast_times_for_month}
\alias{forecast_times_for_month}
\title{Return forecast reference times in a particular calendar month}
\usage{
forecast_times_for_month(svals, month)
}
\arguments{
\item{svals}{vector of months since January 1920}

\item{month}{calendar month from 1 to 12}
}
\value{
subset of \code{svals}
}
\description{
Return forecast reference times in a particular calendar month
}
