% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_raster.R
\name{generate_raster}
\alias{generate_raster}
\title{Generate a netCDF populated with data from a generator}
\usage{
generate_raster(
  fname,
  varnames = c("data"),
  generator = stats::runif,
  nrow = 18,
  ncol = 36,
  extent = c(-180, 180, -90, 90),
  attrs = list(),
  append = FALSE
)
}
\arguments{
\item{fname}{of the generated netCDF}

\item{varnames}{character vector of variable names
to write to \code{fname}}

\item{generator}{a function that will be called with
no arguments to produce values to
populate the variables}

\item{nrow}{number of rows in the generated
netCDF}

\item{ncol}{number of columns in the generated
netCDF}

\item{extent}{vector of (xmin, xmax, ymin, ymax)}

\item{attrs}{List of attributes to associate with the file,
or with each variable.  Each attribute is
described by a list, with the following
properties:
\describe{
\item{var}{Variable with which the attribute
is associated, or NULL for a global attribute}
\item{key}{Name of the attribute}
\item{val}{Value of the attribute}
}}

\item{append}{Determines if we should add variables to an existing
file, if present.}
}
\description{
Generate a netCDF populated with data from a generator
}
