% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_command.R
\name{get_command}
\alias{get_command}
\title{Get the command line string used to execute the current program}
\usage{
get_command()
}
\description{
Get the command line string used to execute the current program
}
