% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{get_extent}
\alias{get_extent}
\title{Compute the extent of a netCDF file. Doesn't handle longitude rotation.}
\usage{
get_extent(cdf)
}
\arguments{
\item{cdf}{object of class \code{ncdf4}}
}
\value{
vector of (xmin, xmax, ymin, ymax)
}
\description{
Compute the extent of a netCDF file. Doesn't handle longitude rotation.
}
