% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{get_latvar}
\alias{get_latvar}
\title{Determine the name of the latitude variable in a netCDF}
\usage{
get_latvar(cdf)
}
\arguments{
\item{cdf}{object of class \code{ncdf4}}
}
\value{
name of the latitude variable
}
\description{
Determine the name of the latitude variable in a netCDF
}
