% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{get_lonvar}
\alias{get_lonvar}
\title{Determine the name of the longitude variable in a netCDF}
\usage{
get_lonvar(cdf)
}
\arguments{
\item{cdf}{object of class \code{ncdf4}}
}
\value{
name of the longitude variable
}
\description{
Determine the name of the longitude variable in a netCDF
}
