% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{info}
\alias{info}
\title{Generate a log message at the \code{INFO} level}
\usage{
info(...)
}
\arguments{
\item{...}{Any number of string-convertible objects}
}
\description{
Generate a log message at the \code{INFO} level
}
