% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{infof}
\alias{infof}
\title{Format and generate a log message at the \code{INFO} level}
\usage{
infof(msg, ...)
}
\arguments{
\item{msg}{A message to log. May be a format string.}

\item{...}{Arguments to fill placeholders in format string}
}
\description{
Format and generate a log message at the \code{INFO} level
}
