% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vardef.R
\name{is.wsim.io.var}
\alias{is.wsim.io.var}
\title{Check if an object is a \code{wsim.io.var}}
\usage{
is.wsim.io.var(v)
}
\arguments{
\item{v}{a thing to test}
}
\value{
true if \code{v} is a \code{wsim.io.var}
}
\description{
Check if an object is a \code{wsim.io.var}
}
