% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vardef.R
\name{is.wsim.io.vardef}
\alias{is.wsim.io.vardef}
\title{Check if an object is a \code{wsim.io.vardef}}
\usage{
is.wsim.io.vardef(v)
}
\arguments{
\item{v}{a thing to test}
}
\value{
true if \code{v} is a \code{wsim.io.vardef}
}
\description{
Check if an object is a \code{wsim.io.vardef}
}
