% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_precl.R
\name{is_byte_defined}
\alias{is_byte_defined}
\title{Check is a given pixel is defined in a file accessible through a URL}
\usage{
is_byte_defined(url, position, size, endian, nodata_value)
}
\arguments{
\item{url}{url to probe}

\item{position}{intial position to read from \code{url} (0-indexed)}

\item{size}{number of bytes to read}

\item{endian}{endianness of data, e.g. \code{"little"}}

\item{nodata_value}{value indicating missing data, e.g. \code{-999}}
}
\value{
TRUE if the byte is defined, FALSE otherwise
}
\description{
Check is a given pixel is defined in a file accessible through a URL
}
