% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_precl.R
\name{is_precl_available}
\alias{is_precl_available}
\title{Return TRUE if PREC/L data is available for a given year/month}
\usage{
is_precl_available(year, month)
}
\arguments{
\item{year}{year to test}

\item{month}{month to test}
}
\value{
TRUE if the data is available by FTP
}
\description{
Return TRUE if PREC/L data is available for a given year/month
}
