% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_seq.R
\name{lat_seq}
\alias{lat_seq}
\title{Compute a sequence of latitude values (N to S)}
\usage{
lat_seq(extent, dims)
}
\arguments{
\item{extent}{of the form \code{xmin, xmax, ymin, ymax}}

\item{dims}{matrix dimensions}
}
\value{
sequence of latitudes
}
\description{
Compute a sequence of latitude values (N to S)
}
