% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmme.R
\name{lead_months}
\alias{lead_months}
\title{Compute the number of lead months between a forecast reference time
and a target month}
\usage{
lead_months(reftime, target)
}
\arguments{
\item{reftime}{the YYYYMM when the forecast was generated}

\item{target}{the YYYYMM targeted by the forecast}
}
\value{
the number of lead months (value of the \code{L} dimension)
}
\description{
Compute the number of lead months between a forecast reference time
and a target month
}
