% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging_init}
\alias{logging_init}
\title{Initialize logging functionality}
\usage{
logging_init(tool_name)
}
\arguments{
\item{tool_name}{Name of the program, to be used in
formatting log messages}
}
\description{
Initializes logging at the level set in the environment
variable \code{WSIM_LOGGING}. Calling this function be
the first thing any WSIM program does.
}
