% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_seq.R
\name{lon_seq}
\alias{lon_seq}
\title{Compute a sequence of longitude values (W to E)}
\usage{
lon_seq(extent, dims)
}
\arguments{
\item{extent}{of the form \code{xmin, xmax, ymin, ymax}}

\item{dims}{matrix dimensions}
}
\value{
sequence of longitudes
}
\description{
Compute a sequence of longitude values (W to E)
}
