% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_netcdf_dims.R
\name{make_netcdf_dims}
\alias{make_netcdf_dims}
\title{Create netCDF dimensions}
\usage{
make_netcdf_dims(vars, extent, ids, extra_dims, unlimited_dims = NULL)
}
\arguments{
\item{vars}{List of matrices containing values to write,
or a 3D with dimension names on the third
dimension.}

\item{extent}{vector of (xmin, xmax, ymin, ymax)}

\item{ids}{vector of integer or character IDs}

\item{extra_dims}{Optional list of extra dimensions (excluding
'id' or spatial dimensions). Specified as a list
with dimension names as keys and a vector of
acceptable values as values, e.g.
\code{extra_dims=list(crop=c('maize', 'rice'), season=c('spring', 'fall'))}}

\item{unlimited_dims}{character vector of dimension names which should
be created as unlimited}
}
\value{
list of \code{ncdim4} objects
}
\description{
Create netCDF dimensions
}
