% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vardef.R
\name{make_var}
\alias{make_var}
\title{Create a \code{wsim.io.var}}
\usage{
make_var(var_in, var_out = NULL, transforms = as.character(c()))
}
\arguments{
\item{var_in}{name of the variable to read}

\item{var_out}{optional name to which the variable
should be changed. If left as the default
NULL, \code{var_out} will be the same as
\code{var_in}.}

\item{transforms}{a character vector containing zero or
more text representations of transformations
to be applied to \code{var_in} after it is read.}
}
\description{
A \code{wsim.io.var} is a reference to a variable
to which zero or more transformations may be applied,
and whose name may be changed.
}
