% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vardef.R
\name{make_vardef}
\alias{make_vardef}
\title{Create a \code{wsim.io.vardef}}
\usage{
make_vardef(filename, vars)
}
\arguments{
\item{filename}{filename of vardef}

\item{vars}{list of \code{wsim.io.var}}
}
\value{
a constructed \code{wsim.io.vardef}
}
\description{
A \code{wsim.io.vardef} is a reference to a file
from which one or more variables can be read.
}
