% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmme.R
\name{months_since_jan_1960}
\alias{months_since_jan_1960}
\title{Convert a YYYYMM to months since January 1960}
\usage{
months_since_jan_1960(yearmon)
}
\arguments{
\item{yearmon}{year/month in YYYYMM format}
}
\value{
number of months since January 1960
}
\description{
Convert a YYYYMM to months since January 1960
}
