% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmme.R
\name{nmme_to_halfdeg}
\alias{nmme_to_halfdeg}
\title{Disaggregate a matrix or array from NMME grid to 0.5-degree global grid}
\usage{
nmme_to_halfdeg(m)
}
\arguments{
\item{m}{a 181x360xN array of values from an NMME forecast or climatology.
The values must have been read using \code{\link{read_vars}} so
that the longitude grid originating at -0.5 W has been wrapped to
begin at -179.5 W.}
}
\value{
a 360x720XN array of transformed values
}
\description{
Disaggregate a matrix or array from NMME grid to 0.5-degree global grid
}
