% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_precl.R
\name{open_noaa_precl}
\alias{open_noaa_precl}
\title{Read a gridded 0.5-degree PREC/L binary file}
\usage{
open_noaa_precl(fname, month, layer = 1)
}
\arguments{
\item{fname}{the file name to read}

\item{month}{the month to read}

\item{layer}{layer to retrieve (1 = precipitation rates, 2 = number of stations)}
}
\value{
a an opened file handle, positioned at the given month/layer
}
\description{
PREC/L files are distributed in a custom binary format at the following URL:
ftp://ftp.cpc.ncep.noaa.gov/precip/50yr/gauge/0.5deg/format_bin_lnx/
}
\details{
Each file contains up to 12 months of precipitation rates and gauge counts
}
