% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_attr.R
\name{parse_attr}
\alias{parse_attr}
\title{Parse a metadata attribute destined for a NetCDF file}
\usage{
parse_attr(attr)
}
\arguments{
\item{attr}{text representation of attribute in the
            following format:

            var_name:attr_name=attr_val

            If no var_name is specified, the attribute
            will be interpreted to be a global attribute.

            If no attr_val is specified, it will be set
            to NULL in the returned list.}
}
\value{
a list representation of attribute, with items
        var, key, and val corresponding to var_name, attr_name,
        and attr_val
}
\description{
Parse a metadata attribute destined for a NetCDF file
}
