% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_quantiles.R
\name{parse_quantiles}
\alias{parse_quantiles}
\title{Parse quantile values found in variable names}
\usage{
parse_quantiles(varnames)
}
\arguments{
\item{varnames}{character vector of variable names, possibly
containing entries like \code{RO_q95}}
}
\value{
sorted integer vector of unique parsed quantiles
}
\description{
Parse quantile values found in variable names
}
