% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_stat.R
\name{parse_stat}
\alias{parse_stat}
\title{Parse a statistic provided as a command-line argument}
\usage{
parse_stat(stat)
}
\arguments{
\item{stat}{the argument string}
}
\value{
a parsed \code{wsim.io.stat}
}
\description{
Parses a statistic argument of the form
\code{stat}
or
\code{stat::var1,var2,var3}
}
