% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_transforms.R
\name{perform_transforms}
\alias{perform_transforms}
\title{Apply named transformations to data}
\usage{
perform_transforms(data, transforms)
}
\arguments{
\item{data}{a numeric vector or matrix}

\item{transforms}{a character vector of transformaiton
descriptions}
}
\value{
a transformed version of \code{data}
}
\description{
Apply named transformations to data
}
