% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_precl.R
\name{precl_byte_offset}
\alias{precl_byte_offset}
\title{Return the number of bytes from the start of a PREC/L binary file until a given pixel in a given month}
\usage{
precl_byte_offset(month, pixel, layer)
}
\arguments{
\item{month}{the month to read}

\item{pixel}{pixel in image (1 for start of image)}

\item{layer}{layer to retrieve (1 = precipitation rates, 2 = number of stations)}
}
\value{
offset in bytes
}
\description{
Return the number of bytes from the start of a PREC/L binary file until a given pixel in a given month
}
