% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_precl.R
\name{precl_url}
\alias{precl_url}
\title{Return the FTP url for a PREC/L binary file}
\usage{
precl_url(year)
}
\arguments{
\item{year}{year of observation}
}
\value{
url
}
\description{
Return the FTP url for a PREC/L binary file
}
