% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_blockwise_apply.R
\name{raster_blockwise_apply}
\alias{raster_blockwise_apply}
\title{Apply a function to each cell of a raster, using blockwise I/O}
\usage{
raster_blockwise_apply(
  filename,
  filename_out,
  fun,
  nodata,
  band = 1,
  datatype = "Float32"
)
}
\arguments{
\item{filename}{Filename of raster to read}

\item{filename_out}{Filename of output GeoTIFF}

\item{fun}{Function to apply to each cell of input}

\item{nodata}{NODATA value to use for \code{filename_out}}

\item{band}{The band of \code{filename} to read}

\item{datatype}{Data type of output raster (allowable type
names are defined by the \code{rgdal} package.)}
}
\description{
Apply a function to each cell of a raster, using blockwise I/O
}
