% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_brick_from_cdf.R
\name{read_brick_from_cdf}
\alias{read_brick_from_cdf}
\title{Read multiple variables from a NetCDF file to a RasterBrick}
\usage{
read_brick_from_cdf(fname)
}
\arguments{
\item{fname}{name of NetCDF file}
}
\value{
RasterBrick
}
\description{
Global attributes will be attached to the RasterBrick as
metadata.  Variable attributes are currently ignored.
}
