% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dimension_values.R
\name{read_dimension_values}
\alias{read_dimension_values}
\title{Get the values of the dimensions associated with a variable in a netCDF file}
\usage{
read_dimension_values(vardef, exclude.dims = NULL)
}
\arguments{
\item{vardef}{filename/variable to check}

\item{exclude.dims}{dimensions to ignore in output}
}
\value{
a list with names representing dimension names and values
        representing dimension values
}
\description{
Get the values of the dimensions associated with a variable in a netCDF file
}
