% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fits_from_cdf.R
\name{read_fits_from_cdf}
\alias{read_fits_from_cdf}
\title{Read distribution fit files and store them in a list}
\usage{
read_fits_from_cdf(files)
}
\arguments{
\item{files}{filenames representing netCDF files with distribution parameters}
}
\value{
a list with one 3d array per distribution, keyed on variable name
}
\description{
Throws an error if any of the following conditions apply:
1) Extents of multiple fit files do not match
2) A fit file has an undefined distribution
3) A fit file has an undefined variable
}
