% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_integrated_vars.R
\name{read_integrated_vars}
\alias{read_integrated_vars}
\title{Read variables from several files and return a data frame with a "window" column}
\usage{
read_integrated_vars(vardefs, expected_ids, as.data.frame = TRUE)
}
\arguments{
\item{vardefs}{a list of vardefs}

\item{expected_ids}{expected IDs to check}

\item{as.data.frame}{return results as a data frame?}
}
\value{
variables at provided integration periods
}
\description{
Read variables from several files and return a data frame with a "window" column
}
