% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mon_file.R
\name{read_mon_file}
\alias{read_mon_file}
\title{Read a gridded binary .mon file}
\usage{
read_mon_file(filename, na.value = -999)
}
\arguments{
\item{filename}{filename to read}

\item{na.value}{NODATA value, to be replaced with NA}
}
\value{
360x720 matrix representing the contents of the .mon file
}
\description{
Read a gridded binary file of the format described at
ftp://ftp.cpc.ncep.noaa.gov/wd51yf/global_monthly/gridded_binary/README.txt
}
\details{
The file must be global in extent, with 0.5-degree resolution.
}
