% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmme.R
\name{read_nmme_noaa}
\alias{read_nmme_noaa}
\title{Read a variable from an NMME file as distributed by NOAA}
\usage{
read_nmme_noaa(fname, var, lead_months, member = NULL)
}
\arguments{
\item{fname}{path to netCDF file}

\item{var}{name of variable to read, e.g., 'fcst'}

\item{lead_months}{integer number of lead months, where zero
corresponds to the month when the forecast
was issued}

\item{member}{ensemble member for forecasts, starting with \code{1}}
}
\value{
data
}
\description{
Data will be rotated from 0-360 to -180-180 and disaggregated to
a 0.5-degree global grid.
}
\examples{
\dontrun{
oct_precip_anom <- read_nmme_noaa('/tmp/NASA_GEOS5v2.prate.201909.anom.nc', 'fcst', 1, 3)
oct_precip_clim <- read_nmme_noaa('/tmp/NASA_GEOS5v2.tmp2m.01.mon.clim.nc', 'clim', 1)
}
}
