% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_cpc_global_daily_temp.R
\name{read_noaa_cpc_global_daily_temp}
\alias{read_noaa_cpc_global_daily_temp}
\title{Read a binary file of global daily temperatures from CPC}
\usage{
read_noaa_cpc_global_daily_temp(
  fname,
  varname,
  day_start,
  day_end,
  check_defined = TRUE
)
}
\arguments{
\item{fname}{path to file}

\item{varname}{name of variable to extract (one of \code{tmax}, \code{nmax},
\code{tmin}, \code{nmin})}

\item{day_start}{first day of year to extract}

\item{day_end}{last day of year to extract}

\item{check_defined}{check that values are defined? (entire grid is not NODATA)}
}
\value{
a 360x720xN array of daily values for \code{varname}
}
\description{
Data available from \code{ftp://ftp.cpc.ncep.noaa.gov/precip/PEOPLE/wd52ws/global_temp/}
}
