% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_noaa_precl.R
\name{read_noaa_precl}
\alias{read_noaa_precl}
\title{Read precipitation rates or gauge counts from a file handle}
\usage{
read_noaa_precl(fh, what)
}
\arguments{
\item{fh}{an open file handle pointing to the position to begin reading}

\item{what}{\code{precipitation_rate} or \code{gauge_count}}
}
\description{
Read precipitation rates or gauge counts from a file handle
}
