% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varnames.R
\name{read_varnames}
\alias{read_varnames}
\title{Get the variables available in a given file}
\usage{
read_varnames(fname)
}
\arguments{
\item{fname}{filename to check}
}
\value{
a character vector of variable names
}
\description{
Get the variables available in a given file
}
