% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{read_vars_from_cdf}
\alias{read_vars_from_cdf}
\title{Read all variables from a netCDF file}
\usage{
read_vars_from_cdf(
  vardef,
  vars = as.character(c()),
  offset = NULL,
  count = NULL,
  extra_dims = NULL,
  as.data.frame = FALSE
)
}
\arguments{
\item{vardef}{A variable definition, constructed
              using the following syntax:
              <filename>::<variables_or_bands>
              where \code{<variables_or_bands> } is a comma-separated list
              of band numbers or variable names.

              A limited amount of transformation can be specified by adding
              extra characters to the band number or variable name:
              \itemize{
              \item{Variables can be renamed by appending \code{->new_var_name}}
              \item{Data can be transformed by appending transformations such as
                    \code{@negate@fill0 }}
              }

              A complete example is:
              \code{PETmE_freq_trgt201701.img::1@fill0@negate->Neg_PETmE}

              In this example, band 1 is read from the file. NODATA values
              are replaced with zero, and all values are negated. The values
              are read into a variable called \code{Neg_PETmE}.}

\item{vars}{A character vector of variable names to read.  If NULL (default),
all variables will be read.}

\item{offset}{Specifies dimension-wise (X, Y, ...) offsets
from the upper-left corner of the raster
(xmin, ymax) from which reading should begin.
A value of c(1,1) refers to the corner cell
itself (i.e., there is no offset). This
follows the convention of the \code{ncdf4}
package. If \code{offset} is specified, then
\code{count} must also be specified.}

\item{count}{Specifies dimension-wise (X, Y, ...) numbers of
cells to read, beginning at the origin cell
specified by \code{offset}. A value of \code{-1}
signifies that all values after the origin cell
should be read. If \code{count} is specified, then
\code{origin} must also be specified.}

\item{extra_dims}{list containing names and values of extra dimensions
along which a values at a single point should be extracted, e.g.
\code{extra_dims=list(crop='maize', quantile=0.50)}. It provides
a higher-abstraction alternative to the use of \code{offset} and
\code{count}.}

\item{as.data.frame}{return data in a data frame}
}
\value{
structure described in \code{\link{read_vars}}, or a data frame
}
\description{
Read all variables from a netCDF file
}
