% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_to_cube.R
\name{read_vars_to_cube}
\alias{read_vars_to_cube}
\title{Read multiple variables to a 3D array}
\usage{
read_vars_to_cube(
  vardefs,
  attrs_to_read = as.character(c()),
  offset = NULL,
  count = NULL
)
}
\arguments{
\item{vardefs}{a list or vector of variable definitions
as described in \code{\link{parse_vardef}}}

\item{attrs_to_read}{a vector of global attribute names to be
read from the first variable definition
and attached as attributes to the returned
array.}

\item{offset}{Specifies dimension-wise (X, Y, ...) offsets
from the upper-left corner of the raster
(xmin, ymax) from which reading should begin.
A value of c(1,1) refers to the corner cell
itself (i.e., there is no offset). This
follows the convention of the \code{ncdf4}
package. If \code{offset} is specified, then
\code{count} must also be specified.}

\item{count}{Specifies dimension-wise (X, Y, ...) numbers of
cells to read, beginning at the origin cell
specified by \code{offset}. A value of \code{-1}
signifies that all values after the origin cell
should be read. If \code{count} is specified, then
\code{origin} must also be specified.}
}
\value{
a 3D array.  The dimnames of the third dimension
        will contain the variable names of the inputs, and
       the extent will be attached as an attribute.
}
\description{
Read multiple variables to a 3D array
}
