% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{set_dimnames}
\alias{set_dimnames}
\title{Set names of specified dimension}
\usage{
set_dimnames(x, dim, names)
}
\arguments{
\item{x}{array}

\item{dim}{dimension number (e.g., \code{1}, \code{2})}

\item{names}{character vector of names}
}
\value{
\code{x} with updated dimension names
}
\description{
Set names of specified dimension
}
