% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vars_from_cdf.R
\name{shared_dimensions}
\alias{shared_dimensions}
\title{Given a netCDF file handle and a list of variable names,
return the names of the dimensions that are shared among
all variables.}
\usage{
shared_dimensions(cdf, varnames)
}
\arguments{
\item{cdf}{a ncdf4 file handle}

\item{varnames}{variable names to be included in the scan}
}
\value{
a list of dimension names that are common to all
        variables in \code{varnames}
}
\description{
Given a netCDF file handle and a list of variable names,
return the names of the dimensions that are shared among
all variables.
}
