% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_netcdf_attrs.R
\name{standard_netcdf_attrs}
\alias{standard_netcdf_attrs}
\title{Generate a list of standard attributes to include in netCDF outputs}
\usage{
standard_netcdf_attrs(is_new, is_spatial, existing_history = NULL)
}
\arguments{
\item{is_new}{is the file new?}

\item{is_spatial}{does the file have spatial data?}

\item{existing_history}{optional character vector with an existing history
entry to which we should potential apppend}
}
\value{
list of attributes
}
\description{
Generate a list of standard attributes to include in netCDF outputs
}
