% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{update_attribute}
\alias{update_attribute}
\title{Update an attribute in ncout}
\usage{
update_attribute(ncout, var, key, val, prec)
}
\arguments{
\item{ncout}{file handle for netCDF in which attributes should be
updated}

\item{var}{the name of the variable to which the attribute
is associated (or \code{NULL} for a global attribute)}

\item{key}{the name of the attribute}

\item{val}{the value of the attribute}

\item{prec}{the precision of the attribute (or \code{NULL} to
use the same precision as \code{var})}
}
\description{
Update an attribute in ncout
}
