% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{update_dimnames}
\alias{update_dimnames}
\title{Update names of specified dimension}
\usage{
update_dimnames(x, dim, fun)
}
\arguments{
\item{x}{array}

\item{dim}{dimension number (e.g., \code{1}, \code{2})}

\item{fun}{function called with previous dimensions names
and returning updated dimension names}
}
\value{
\code{x} with updated dimension names
}
\description{
Update names of specified dimension
}
