% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{var_fill}
\alias{var_fill}
\title{Return a fill value to use for the variable named var}
\usage{
var_fill(var, vals, prec)
}
\arguments{
\item{var}{name of the variable}

\item{vals}{list containing data for vars}

\item{prec}{precision argument as described in write_vars_to_cdf}
}
\description{
Return a fill value to use for the variable named var
}
