% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{var_prec}
\alias{var_prec}
\title{Return the data precision for variable named var}
\usage{
var_prec(var, vals, prec)
}
\arguments{
\item{var}{name of the variable}

\item{vals}{list containing data for vars}

\item{prec}{precision argument as described in write_vars_to_cdf}
}
\value{
text representation of precision for variable
}
\description{
Return the data precision for variable named var
}
