% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{warn}
\alias{warn}
\title{Generate a log message at the \code{WARN} level}
\usage{
warn(...)
}
\arguments{
\item{...}{Any number of string-convertible objects}
}
\description{
Generate a log message at the \code{WARN} level
}
