% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{warnf}
\alias{warnf}
\title{Format and generate a log message at the \code{WARN} level}
\usage{
warnf(msg, ...)
}
\arguments{
\item{msg}{A message to log. May be a format string.}

\item{...}{Arguments to fill placeholders in format string}
}
\description{
Format and generate a log message at the \code{WARN} level
}
