% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_layer_to_cdf.R
\name{write_layer_to_cdf}
\alias{write_layer_to_cdf}
\title{Write a RasterLayer to a NetCDF file}
\usage{
write_layer_to_cdf(lyr, filename, varname, attrs = list(), prec = "double")
}
\arguments{
\item{lyr}{RasterLayer to write}

\item{filename}{output filename}

\item{varname}{variable name to use in NetCDF file}

\item{attrs}{list of attributes to attach to the file,
e.g., list(
        list(key='distribution', val='GEV'), # global attribute
        list(var='precipitation', key='units', val='mm)
      )}

\item{prec}{data type for values.  Valid types:
* short
* integer
* float
* double
* char
* byte}
}
\description{
Write a RasterLayer to a NetCDF file
}
