% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stack_to_cdf.R
\name{write_stack_to_cdf}
\alias{write_stack_to_cdf}
\title{Write a RasterStack to a NetCDF file}
\usage{
write_stack_to_cdf(stk, filename, attrs = list(), prec = "double")
}
\arguments{
\item{stk}{RasterStack containing named layers}

\item{filename}{output filename}

\item{attrs}{list of attributes to attach to the file,
e.g., list(
        list(key='distribution', val='GEV'), # global attribute
        list(var='precipitation', key='units', val='mm)
      )}

\item{prec}{data type for values.  Valid types:
* short
* integer
* float
* double
* char
* byte}
}
\description{
Write a RasterStack to a NetCDF file
}
