% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{write_vars_to_cdf}
\alias{write_vars_to_cdf}
\title{Write a list of matrices to a netCDF file.}
\usage{
write_vars_to_cdf(
  vars,
  filename,
  extent = NULL,
  ids = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  extra_dims = NULL,
  write_slice = NULL,
  attrs = list(),
  prec = NULL,
  append = FALSE,
  put_data = TRUE,
  quick_append = FALSE
)
}
\arguments{
\item{vars}{List of matrices containing values to write,
or a 3D with dimension names on the third
dimension.}

\item{filename}{Name of output file}

\item{extent}{vector of (xmin, xmax, ymin, ymax)}

\item{ids}{vector of integer or character IDs}

\item{xmin}{lowest longitude value (left side of cell)}

\item{xmax}{highest longitude value (right side of cell)}

\item{ymin}{lowest latitude value (bottom of cell)}

\item{ymax}{highest latitude value (top of cell)}

\item{extra_dims}{Optional list of extra dimensions (excluding
'id' or spatial dimensions). Specified as a list
with dimension names as keys and a vector of
acceptable values as values, e.g.
\code{extra_dims=list(crop=c('maize', 'rice'), season=c('spring', 'fall'))}}

\item{write_slice}{Optional list used to write a two-dimensional array
(lat, lon) across constant values of other dimensions.
List keys are used to indicate dimension names and values
are used to indicate the constant value for that dimension.}

\item{attrs}{List of attributes to associate with the file,
or with each variable.  Each attribute is
described by a list, with the following
properties:
\describe{
\item{var}{Variable with which the attribute
is associated, or NULL for a global attribute}
\item{key}{Name of the attribute}
\item{val}{Value of the attribute}
}}

\item{prec}{The numerical precision with which \code{vars}
should be written. Can be either a character vector,
in which case the same precision will be used for
all variables, or a list whose names correspond to
the names of \code{vars}, specifying a precision for
each variable. If \code{vars} is \code{NULL}, or a
variable is not specified in a list, a default precision
will be selected depending on the values.
Acceptable precision descriptions include:
\itemize{
\item byte (8-bit signed integer)
\item short (16-bit signed integer)
\item integer (32-bit signed integer)
\item single (32-bit floating point)
\item float (32-bit floating point)
\item double (64-bit floating point)
}}

\item{append}{Determines if we should add variables to an existing
file, if present.}

\item{put_data}{Flag indicating whether to write data contained in
\code{vars}. If \code{FALSE}, only dimensions and
coordinate varibles will be written.}

\item{quick_append}{Flag indicating whether to validate that data
written in an append operation conforms to the
original dimensions of the file. If \code{TRUE},
these checks will not be performed, increasing
write performance.}
}
\description{
The names of the list elements will be used to name the
variable associated with each matrix.  Consistent with the
raster package, it is assumed that matrices are stored as
(lat, lon), with lon increasing and lat decreasing.
}
