% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vars_to_cdf.R
\name{write_wgs84_crs_attributes}
\alias{write_wgs84_crs_attributes}
\title{Write CRS attributes for WGS84}
\usage{
write_wgs84_crs_attributes(ncout, var_names)
}
\arguments{
\item{ncout}{netCDF open for writing}

\item{var_names}{list of variable names to which CRS should be associated}
}
\description{
Write CRS attributes for WGS84
}
