% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsim_version_string.R
\name{wsim_version_string}
\alias{wsim_version_string}
\title{Return a string with the WSIM version and commit hash}
\usage{
wsim_version_string()
}
\description{
Return a string with the WSIM version and commit hash
}
