% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmme.R
\name{yearmon_from_months_since_jan_1960}
\alias{yearmon_from_months_since_jan_1960}
\title{Convert months since January 1960 to YYYYMM}
\usage{
yearmon_from_months_since_jan_1960(s)
}
\arguments{
\item{s}{number of months since January 1960}
}
\value{
year and month in YYYYMM format
}
\description{
Convert months since January 1960 to YYYYMM
}
